[#174] Refactor eslint config and create rule for i18n keys [#174] Fix code with ts imports
