+++ src-ds/docs/components/docs-page.tsx
@@ -5,7 +5,7 @@ import { SlotsTable } from '@ds/docs/components/_partials/slots-table'
 import { useDocsPlayground } from '@ds/docs/components/docs-playground-provider'
 import { IconButton, MaximizeSvg, MinimizeSvg } from '@ds/release'
 import { ReactNode, useCallback, useEffect, useRef, useState } from 'react'
-import { renderHtml } from '../utilities/docs.ts'
+import { renderHtml } from '../utilities/docs'
 import { toggleControls } from '../utilities/storybook'
 import { DocsHeader } from './docs-header'
 import { DocsPlayground } from './docs-playground'
