+++ src-ds/docs/stories/0-home/1-changelog.story.tsx
@@ -1,7 +1,7 @@
-import { CHANGELOG, LogEntry } from '@ds/changelog.ts'
+import { CHANGELOG, LogEntry } from '@ds/changelog'
 import { DocsPage } from '@ds/docs/components/docs-page'
 import '@ds/docs/setup'
-import { renderHtml } from '@ds/docs/utilities/docs.ts'
+import { renderHtml } from '@ds/docs/utilities/docs'
 import type { StoryObj } from '@storybook/react'
 
 export const story: StoryObj = {}
