+++ src-ds/docs/stories/4-components/modal.story.tsx
@@ -1,5 +1,5 @@
-import { DocsPage } from '@ds/docs/components/docs-page.tsx'
-import { createArgTypes } from '@ds/docs/setup.ts'
+import { DocsPage } from '@ds/docs/components/docs-page'
+import { createArgTypes } from '@ds/docs/setup'
 import { Button, Modal, ModalProps } from '@ds/release'
 import { action } from '@storybook/addon-actions'
 import type { Meta, StoryObj } from '@storybook/react'
